Committing part of the code refactor:
 - CountVector now holds a LIST of event dictionaries. These dictionaries store events separately for different tiles. During the DP,
they are folded into each other to allow the merging behavior to compute actual frequencies, but this new structure also allows access
to the set of possible tiles, and the events for those tiles, even after the DP is done.

 - CountVector has a few new methods - mostly ways to fold the events into each other, and to manipulate the events for a CVS

 - DLCeventscape now has two new flags, which allows the user to specify whether to calculate events, and whether to draw the regions
of optimality to the screen.

Some things have yet to be changed, namely recon.py - dlcpar and dlcscape were broken due to changes while trying to get
reconeventscape to work.